Split into files and simplify main.go
