+++ helper.go
@@ -1,0 +1,8 @@
+package main
+
+import "os"
+
+func fileExists(filename string) bool {
+	_, err := os.Stat(filename)
+	return !os.IsNotExist(err)
+}
